Add initial working version of geonode's geoserver in Minikube.
Note that this example was originally developed in the directory:
/home/tarter/GIS/geonode/K8
Some of the internals within the yaml will need adjustment for other locations.
The Dockerized version of geoserver used is from:
https://hub.docker.com/r/geonode/geoserver
This solution is also dependent on data downloaded from:
https://build.geo-solutions.it/geonode/geoserver/latest/
IMPORTANT NOTE: This solution counts on the fact that Minikube makes the contents of a user's home directory visible as a persistent volume, at least on ubuntu.